+++ RogueLike/TileMap.cpp
@@ -7,7 +7,7 @@ TileMap::TileMap()
 }
 
 bool TileMap::load(const char* filename, const Map& map, 
-				   int tile_size, int width, int height)
+				   size_t tile_size, size_t width, size_t height)
 {
 	cout << "Loading tileset from file:" << filename << endl;
 	if(!tileset_.loadFromFile(filename))
